refactor: rename testing folder to test (#56)
