+++ test/root_command_flags_test.go
@@ -6,7 +6,7 @@ import (
 	qt "github.com/frankban/quicktest"
 
 	"github.com/chiselstrike/libsql-shell/cmd"
-	"github.com/chiselstrike/libsql-shell/testing/utils"
+	"github.com/chiselstrike/libsql-shell/test/utils"
 )
 
 func TestRootCommandFlags_WhenAllFlagsAreProvided_ExpectSQLStatementsExecutedWithoutError(t *testing.T) {
